Selective Display Ellipsis Proof-of-Concept
Coercing `beacon-dec' into playing well with this was _way_ hairier
than I was expecting it to be, and I stayed up too late to get
it (mostly) working (something funny is still happening with the
foreground of the before-string, problem for another day). Definitely
not my best work, but it's well past time for a POC.

Haven't checked to see if it works in TUI frames, but I'm more
optimistic about that than I was earlier, knowing now that I can set
that width in `beacon--selective-display-set-face-at-point' to zero.